BackendFastAPI/requirements.txt
fastapi
uvicorn
motor
pydantic
PyJWT
python-dotenv
pytest
pytest-asyncio
httpx
bson
pymongo